refactor: VideoCardView and ViewModel
