+++ src/Spice86/Views/VideoCardView.axaml.cs
@@ -1,0 +1,11 @@
+using Avalonia;
+using Avalonia.Controls;
+using Avalonia.Markup.Xaml;
+
+namespace Spice86.Views;
+
+public partial class VideoCardView : UserControl {
+    public VideoCardView() {
+        InitializeComponent();
+    }
+}
